Upload video endpoint

+++ backend/backend/Models/Video.cs
@@ -14,7 +14,7 @@ namespace backend.Models
 
         [Required]
         public string Title { get; set; }
-        public double Size { get; set; }
+        public long Size { get; set; }
 
         [Required]
         public DateTime UploadDate { get; set; }
